requirements.txt
numpy==1.26.4      # <-- pin!
opencv-python==4.10.0.82
pyautogui==0.9.54
pyyaml==6.0.1
pillow==10.3.0